+++ qase-javascript-commons/src/models/index.ts
@@ -2,6 +2,7 @@ export { TestResultType } from './test-result';
 export type { Relation, Suite, SuiteData, TestopsProjectMapping } from './test-result';
 export { TestExecution, TestStatusEnum } from './test-execution';
 export { TestStepType, StepType } from './test-step';
+export type { StepRequestData, StepTextData, StepGherkinData } from './step-data';
 export { StepStatusEnum } from './step-execution';
 export type { Attachment } from './attachment';
 export type { Report } from './report';